Fix message shown for geo-location on error.
The get geo-location initiated by a single click shows an error message on failure to send message to the phone app. If the phone had been running and displaying messages on the Pebble, an error message on the Pebble was truncated if the phone app stopped running. The truncation was due to the scrolling layer not being set properly for the error message.