Project renamed GeoTrail
In Cloud Pebble IDE PebbleMsg project renamed to GeoTrail because want the name in Pebble Store to be GeoTrail.

In github, the project name is still PebbleMsg -- no need to add another github project. The version is still 1.1.009. The only change has been to show GeoTrail instead of MyMsg when Pebble app starts.